requirements.txt
fastapi
uvicorn
python-dotenv
httpx
langchain
langchain-google-genai
faiss-cpu
pydantic
python-multipart